label data, visualize, and resize

+++ download_extract.py
@@ -2,9 +2,10 @@ import urllib.request
 import os
 import numpy as np
 import time
+import cv2
+import keras
+import sys
 
-test_split = 0.2
-val_split = 0.2
 
 def get_files(path):
     files = []
@@ -13,17 +14,22 @@ def get_files(path):
             if file.endswith(".tif"):
                 files.append(os.path.join(r, file))
     return files
-
-os.system("wget http://www.svcl.ucsd.edu/projects/anomaly/UCSD_Anomaly_Dataset.tar.gz -O compressed.tar.gz")
-os.mkdir("dataraw")
-os.system("tar -xf compressed.tar.gz -C dataraw")
-os.remove("compressed.tar.gz")
-os.mkdir("images")
+if False:
+    os.system("wget http://www.svcl.ucsd.edu/projects/anomaly/UCSD_Anomaly_Dataset.tar.gz -O compressed.tar.gz")
+    os.mkdir("dataraw")
+    os.system("tar -xf compressed.tar.gz -C dataraw")
+    os.remove("compressed.tar.gz")
+    os.mkdir("images")
 files = get_files("dataraw")
 
 for i in range(len(files)):
     print(f"{i+1}/{len(files)}")
     filenamejpg = str(i) + ".jpg"
-    os.system(f"ffmpeg -i {files[i]} images/{filenamejpg}")
+    try:
+        img = keras.utils.load_img(files[i], target_size=(224, 224),keep_aspect_ratio=True,color_mode="grayscale")
+        img.save("images/" + filenamejpg)
+    except:
+        print("Error with file",files[i])
+        continue
 os.system("rm -r dataraw")
 print("Done!")
